replay trait implementation ref fix
