+++ Cargo.toml
@@ -17,7 +17,7 @@ members = [
 resolver = "2"
 
 [workspace.package]
-version         = "0.26.0-rc12"
+version         = "0.26.1"
 edition         = "2021"
 readme          = "README.md"
 repository      = "https://github.com/MegaRockLabs/smart-account-auth"
@@ -41,12 +41,12 @@ saa-proto-solana        = { version = "0.26.0", path = "packages/macros-proto/so
 saa-proto-substrate     = { version = "0.26.0", path = "packages/macros-proto/substrate" }
 
 saa-schema              = { version = "0.26.0", path = "packages/schema" }
-saa-common              = { version = "0.26.0-rc12", path = "packages/common", default-features = false }
-saa-crypto              = { version = "0.26.0-rc12", path = "packages/crypto" }
-saa-auth                = { version = "0.26.0-rc12", path = "packages/auth" }
-saa-curves              = { version = "0.26.1", path = "packages/curves" }
-saa-passkeys            = { version = "0.26.0-rc2", path = "packages/passkeys" }
-smart-account-auth      = { version = "0.26.0-rc12", path = "packages/bundle", default-features = false }
+saa-common              = { version = "0.26.1", path = "packages/common", default-features = false }
+saa-crypto              = { version = "0.26.1", path = "packages/crypto" }
+saa-auth                = { version = "0.26.1", path = "packages/auth" }
+saa-curves              = { version = "0.26.3", path = "packages/curves" }
+saa-passkeys            = { version = "0.26.1", path = "packages/passkeys" }
+smart-account-auth      = { version = "0.26.1", path = "packages/bundle", default-features = false }
