+++ packages/common/Cargo.toml
@@ -74,7 +74,6 @@ wasm       = [
     "dep:schemars",
     "saa-schema/cosmwasm",
 ]
-optimise   = []
 replay     = []
 types      = [
     "dep:schemars",
