+++ packages/crypto/Cargo.toml
@@ -30,7 +30,6 @@ serde               = { workspace = true }
 default     =   ["std"]
 utils       =   []
 replay      =   ["saa-common/replay"]
-optimise    =   ["saa-common/optimise"]
 
 
 std         =   ["saa-common/std",                     "p256?/std"]
