+++ packages/curves/Cargo.toml
@@ -5,7 +5,7 @@ edition     = { workspace = true }
 authors     = { workspace = true }
 license     = { workspace = true }
 repository  = { workspace = true }
-version     = "0.26.1"
+version     = "0.26.3"
 
 
 [dependencies]
