+++ packages/passkeys/Cargo.toml
@@ -5,7 +5,7 @@ edition     = { workspace = true }
 authors     = { workspace = true }
 license     = { workspace = true }
 repository  = { workspace = true }
-version     = "0.26.0-rc2"
+version     = { workspace = true }
 
 
 [dependencies]
