+++ scripts/publish.sh
@@ -12,9 +12,11 @@ cd packages/common && cargo publish && cd ../..
 sleep 10
 cd packages/crypto && cargo publish && cd ../..
 sleep 10
-cd packages/curves && cargo publish && cd ../..
-sleep 10
+# cd packages/curves && cargo publish && cd ../..
+# sleep 10
 cd packages/auth && cargo publish && cd ../..
+sleep 10
+cd packages/passkeys && cargo publish && cd ../..
 sleep 10
 cd packages/bundle && cargo publish && cd ../..
 echo "All packages published"
